docker-app-master/Notification_Api_Docker/requirements.txt
astroid==2.3.3
asyncpg==0.20.1
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
databases==0.2.6
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.54.1
fastapi-users==0.8.0
fastapi-login==1.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.3.31
httptools==0.1.1
httpx==0.11.1
httpx-oauth==0.2.1
hyperframe==5.2.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
passlib==1.7.2
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.2
python-dotenv==0.12.0
python-multipart==0.0.5
redis==3.4.1
requests==2.23.0
rfc3986==1.3.2
six==1.14.0
sniffio==1.1.0
SQLAlchemy==1.3.15
starlette==0.13.2
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
wrapt==1.11.2
zipp==3.1.0
